Exercício 3/pyproject.toml
[tool.poetry]
name = "Exercício 4"
version = "0.1.0"
description = ""
authors = ["Yuri-Almir-Pinto <yurialmirp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.1.1"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"
qrcode = "^7.4.2"
pillow = "^10.2.0"
pyzbar = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
